+++ pyannote/audio/pipelines/resegmentation.py
@@ -95,7 +95,7 @@ class Resegmentation(SpeakerDiarizationMixin, Pipeline):
 
         model: Model = get_model(segmentation, use_auth_token=use_auth_token)
         self._segmentation = Inference(model)
-        self._frames = self._segmentation.model.output_frames
+        self._frames = self._segmentation.model.example_output.frames
 
         self._audio = model.audio